Merging upstream repo to local

+++ src/com/gainsight/sfdc/sfWidgets/accWidget/tests/AccWidget_CockpitTests.java
@@ -339,8 +339,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		        	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		        	task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		        	task.setFromCustomer360orWidgets(true);
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+					task.setFromCustomer360orWidgets(true);
 		        	}
 		       accWfPage  = accWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
 		       for(Task task : tasks) {
@@ -366,8 +366,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		        	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		        	task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		        	task.setAssignee(users[i]); if(++i >=5) i=0;
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+					task.setAssignee(users[i]); if(++i >=5) i=0;
 		        	task.setFromCustomer360orWidgets(true);
 		        	}
 
@@ -393,8 +393,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		        	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+					task.setFromCustomer360orWidgets(true);
 		        	}
 		        
 		        //Applying Playbook and verifying tasks
@@ -409,8 +409,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		          	if(task.getAssignee()==null) {
 		                task.setAssignee(sfdcInfo.getUserFullName());
 		            }
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		          	}
 		       cta.setDueDate(getHighestTaskDate(tasks));
 		       accWfPage = accWfPage.applyPlayBook(cta, testData.get("UpdatedPlaybook"), updatedTasks,false);
@@ -438,8 +438,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		        for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+					task.setFromCustomer360orWidgets(true);
 		        	}
 
 		       accWfPage  = accWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
@@ -463,8 +463,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		        for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+					task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+					task.setFromCustomer360orWidgets(true);
 		        	}
 		        
 		        //Applying Playbook and verifying tasks
@@ -477,8 +477,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		       ArrayList<Task> updatedTasks = getTaskFromSFDC(testData.get("UpdatedPlaybook"));
 		       for(Task task : updatedTasks) {
 		          	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		           task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		           task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		          	}
 		       cta.setDueDate(getHighestTaskDate(tasks));
 		       accWfPage = accWfPage.applyPlayBook(cta, testData.get("UpdatedPlaybook"), updatedTasks,false);
@@ -506,8 +506,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		       for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		            task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		            task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		        	}
 		        
 		       accWfPage = accWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
@@ -530,8 +530,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		       ArrayList<Task> tasks  = getTaskFromSFDC(testData.get("Playbook"));
 		       for(Task task : tasks) {
 		        	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		           task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		           task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		        	}
 		        //Applying Playbook and verifying tasks
 		       accWfPage  = accWfPage.applyPlayBook(cta, testData.get("Playbook"), tasks,true);
@@ -543,8 +543,8 @@ public class AccWidget_CockpitTests extends WorkflowSetup {
 		       ArrayList<Task> updatedTasks = getTaskFromSFDC(testData.get("UpdatedPlaybook"));
 		       for(Task task : updatedTasks) {
 		          	if(task.getAssignee()==null) task.setAssignee(sfdcInfo.getUserFullName());
-		           task.setDate(getTaskDateForPlaybook(Integer.valueOf(task.getDate())));
-		           task.setFromCustomer360orWidgets(true);
+				   task.setDate(getTaskDateForPlaybook(Double.valueOf(task.getDate())));
+				   task.setFromCustomer360orWidgets(true);
 		          	}
 		       cta.setDueDate(getHighestTaskDate(tasks));
 		       accWfPage = accWfPage.applyPlayBook(cta, testData.get("UpdatedPlaybook"), updatedTasks,false);
